Smooth transparent<->opaque transition
I went for a barely-there 200ms animation (configurable at the top of
extension.js), aiming for sleek and unobtrusive rather than eye-candy.

The tween can potentially conflict with other animations so let's try
to minimise the damage by only cancelling tweens when we know we're
animating. Let's see how it works out in the real world.